\name{Envir}
\alias{Envir}
\title{
A temporary environment to stock data and objects
}
\description{
The function create an environment where the data, arguments 
 and objects that are used between the differents function of 
 the package will be stock for better exchange processes.
}
\usage{
Envir()
}
\details{
Objects passed through the environment 'Envir' are called in the
 other function as Envir$objects
}
