\name{selectdirectory}
\alias{selectdirectory}
\title{
Saved path selection
}
\description{
Allow to chose the directory where results (.csv and .png files) will be
 saved.
}
\usage{
selectdirectory()
}
\details{
It select the main save directory; the package will create appropiate sub-folder
as function of selected parameters, statsitics, methods etc.  Then you will be 
able to perform successive analyses wihtout overwriting the previous results.
}
