\name{SRNDunkerque}
\docType{data}
\alias{SRNDunkerque}
\title{ Coastal survey near the Gravelines power plant form 1995 to 2010 }
\description{
Variation in temperature, salinity and chlorophyll-a concentration (microg/l) monthly measured between 1995 and 2010
at three different stations distributed onshore to offshore (North See) near the city of Dunkerque 
(north of France) for the SRN monitoring program (Ifremer).
This database contain many missing values.
}

\format{ A data.frame (TXT) containing 1561 measurments of temperature, salinity and chlorophyll-a concentration  }
\source{ The Ifremer QUADRIGE_2 meta-database }
\keyword{datasets}

