\name{TTAinterface}
\alias{TTAinterface}
\title{
Graphic Interface For Temporal Trend Analysis
}
\description{
A friendly user graphic interface to perform temporal trend analysis.
 The interface offer multiple options to select parameters and build time series 
 that the user can follow step by step. Some options are selected by default 
 to let the hurry user to do really quick analysis. 
 Some diagnosic tools are also present.
}

\usage{
TTAinterface()
}

\value{
Results are saved in .txt files or .png figures in the desire directory (see selectdirectory).
 See 'FULLoption' values fore more details.
}

\author{
David Devreker
}

