\name{TTAinterface-package}
\alias{TTAinterface-package}
\alias{Temporal Trend Analysis interface package}
\docType{package}
\title{
Interface Package for Temporal Trend Analysis
}
\description{
A friendly interface to perform Temporal Trend Analyses (MannKendall tests).
Just follow the successive step from the data frmatting to the results sorting. 
}
\details{
\tabular{ll}{
Package: \tab TTAinterface\cr
Type: \tab Package\cr
Version: \tab 1.5.9\cr
Date: \tab 2021-05-27\cr
License: \tab GPL (>=2)\cr
 }
}

\author{
David Devreker, Alain Lefebvre\cr
Maintainer: <david.devreker@ifremer.fr>
}

\references{ Devreker, D. and Lefebvre, A. (2014), TTAinterfaceTrendAnalysis: An R GUI for routine temporal trend analysis and diagnostics. \emph{Journal of Oceanography, Research and Data}, 1(7), 1-18. }

\keyword{GUI} 
\keyword{temporal trend analyses}
\keyword{time series regularisation}
