\name{TTAinterface-package}
\alias{TTAinterface-package}
\alias{Temporal Trend Analysis interface package}
\docType{package}
\title{
Interface Package for Temporal Trend Analysis
}
\description{
A friendly interface to perform Temporal Trend Analyses (MannKendall tests).
Just follow the successive step from the data frmatting to the results sorting. 
}
\details{
\tabular{ll}{
Package: \tab TTAinterface\cr
Type: \tab Package\cr
Version: \tab 1.01\cr
Date: \tab 2012-03-12\cr
License: \tab GPL (>=2)\cr
 }
}

\author{
David Devreker, Alain Lefebvre\cr
Maintainer: <Alain.Lefebvre@ifremer.fr>
}


\keyword{ 
package 
temporal trend analyses
time series regularization 
}
